rename: Hashtag, HashtagType 파일을 /room/facility 경로로 이동, AccommodationType 파일을 /domain/accommodation 경로로 이동
